requirements.txt
flask
flask-smorest
python-dotenv
PyYAML
requests